+++ ccl-maven-plugin/src/main/java/com/cerner/ccl/testing/maven/ccl/mojo/BaseCclMojo.java
@@ -219,8 +219,8 @@ public abstract class BaseCclMojo extends AbstractMojo {
     /**
      * A regular expression that will match the back end operating system prompt for the configured user. This is used
      * to recognize when the back end operating system has finished processing and is ready for another command. The
-     * following value is constructed by default: $ccl-domainUsername(:$ccl-environment)?@$ccl-host:[^&gt;~#]*[&gt;~#]
-     * or $ccl-domainUsername(:\w*)?@$ccl-host:[^&gt;~#]*[&gt;~#] in case $ccl-host is not specified.
+     * following value is constructed by default: $ccl-domainUsername(:$ccl-environment)?@$ccl-host:[^&gt;~#$]*[&gt;~#$]
+     * or $ccl-domainUsername(:\w*)?@$ccl-host:[^&gt;~#$]*[&gt;~#$] in case $ccl-host is not specified.
      *
      * @since 3.0
      */