+++ whitenoise/whitenoise-maven-plugin/src/main/java/com/cerner/ccl/analysis/mojo/WhitenoiseMojo.java
@@ -182,7 +182,8 @@ public class WhitenoiseMojo extends AbstractMavenReport {
     /**
      * A regular expression that will match the back end operating system prompt for the configured user. This is used
      * to recognize when the back end operating system has finished processing and is ready for another command The
-     * following value is constructed by default: "$ccl-domainUsername(:ccl-environment)?@$ccl-host:[^&gt;~#]*[&gt;~#]"
+     * following value is constructed by default:
+     * "$ccl-domainUsername(:ccl-environment)?@$ccl-host:[^&gt;~#$]*[&gt;~#$]"
      *
      * @since 2.0
      */